requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
heroku==0.1.4
pytube3==9.6.4
requests==2.23.0
requests-toolbelt==0.9.1
youtube-transcript-api==0.3.1